+++ docs/src/markdown/changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.1.0a2
 
+- **NEW**: Add option to group consecutive Python comments.
 - **FIX**: Properly return error.
 - **FIX**: Only retry with default encoding if exception thrown was a `UnicodeDecodeError`.